python/tests: completed meta2 tests on contents
Know TODO: SPARE requests are not tested yet.
